Move controller files to /controllers
