+++ pages/api/controllers/transactions/edit.js
@@ -1,4 +1,4 @@
-const db = require("../db");
+const db = require("../../db");
 
 export default async (req, res) => {
 	// TODO
